chore: clean up again
